custom-Plot-classes: Added a threading.Event to delay the return of results to lyse if the user requests it.
This allows the user to use the Plot subclass to interact with the data, and save additional results to HDF5 file via custom actions on their toolbar (which could be the result of user interaction with the plot) and then return the results to lyse (for example either on a button click, or after a timeout, or whatever you like!)

To use, call `lyse.delay_results_return()` at the end of your script.
Results will then only be returned to lyse (so lyse will be stuck waiting for your analysis script to complete) until `lyse.delay_event.set()` is called (through some event ultimately processed by the Qt event loop).