bugfix: Fix for issue #52. Remove deprecated convert_objects() call.
In current versions of pandas and lyse, in normal use the majority of dataframe
columns that can be are already specific datatypes and not 'object', so there is
limited performance improvement to this.

Some colums that were temporarily a mixed datatype (due to the analysis
results being saved by single shot routines changing type) will remain dtype
object even if all the entries are later of the same type, which is not ideal.

This should be rare and so not likely to be a performance problem, which was the
original purpose of adding convert_objects. But these columns may appear differently
from the perspective of multishot routines, and some code may not be able to handle
them. However we will see empirically if this causes any problems, as perhaps
all relevant code can in fact handle the mixed datatypes fine, so we shouldn't
fix what isn't broken.